+++ ackpine-core/src/main/kotlin/ru/solrudev/ackpine/Ackpine.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023 Ilya Fomichev
+ * Copyright (C) 2023-2024 Ilya Fomichev
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,6 +25,7 @@ import androidx.core.app.NotificationChannelCompat
 import androidx.core.app.NotificationManagerCompat
 import ru.solrudev.ackpine.core.R
 import ru.solrudev.ackpine.exceptions.AckpineReinitializeException
+import ru.solrudev.ackpine.session.parameters.Confirmation.DEFERRED
 
 /**
  * A library providing consistent APIs for installing and uninstalling apps on an Android device.
@@ -39,6 +40,34 @@ public object Ackpine {
 	private val configurationChangesCallback = object : ComponentCallbacks {
 		override fun onConfigurationChanged(newConfig: Configuration) = createNotificationChannel()
 		override fun onLowMemory() {}
+	}
+
+	/**
+	 * Deletes previously created Ackpine's notification channel from application's notification settings.
+	 *
+	 * Use this **only** in conjunction with disabling automatic Ackpine initialization and if [DEFERRED] confirmation
+	 * is never used in the app.
+	 *
+	 * To disable Ackpine initializer, add the following lines to the app's `AndroidManifest.xml`:
+	 *
+	 * ```xml
+	 * <provider
+	 *     android:name="androidx.startup.InitializationProvider"
+	 *     android:authorities="${applicationId}.androidx-startup"
+	 *     android:exported="false"
+	 *     tools:node="merge">
+	 *     <meta-data
+	 *         android:name="ru.solrudev.ackpine.AckpineInitializer"
+	 *         tools:node="remove" />
+	 * </provider>
+	 * ```
+	 */
+	@JvmStatic
+	public fun deleteNotificationChannel(context: Context) {
+		val applicationContext = context.applicationContext
+		applicationContext.unregisterComponentCallbacks(configurationChangesCallback)
+		val channelId = applicationContext.getString(R.string.ackpine_notification_channel_id)
+		NotificationManagerCompat.from(applicationContext).deleteNotificationChannel(channelId)
 	}
 
 	@JvmSynthetic
@@ -59,11 +88,11 @@ public object Ackpine {
 			return
 		}
 		val applicationContext = context ?: return
-		val channelIdString = applicationContext.getString(R.string.ackpine_notification_channel_id)
+		val channelId = applicationContext.getString(R.string.ackpine_notification_channel_id)
 		val channelName = applicationContext.getString(R.string.ackpine_notification_channel_name)
 		val channelDescription = applicationContext.getString(R.string.ackpine_notification_channel_description)
 		val importance = NotificationManager.IMPORTANCE_HIGH
-		val channel = NotificationChannelCompat.Builder(channelIdString, importance)
+		val channel = NotificationChannelCompat.Builder(channelId, importance)
 			.setName(channelName)
 			.setDescription(channelDescription)
 			.setVibrationEnabled(true)